kartons/cape_submitter/requirements.txt
karton-core>=5.0.0
mwdblib>=4.0.0
requests>=2.28.0
